+++ src/api/ExchangeIssuanceAPI.ts
@@ -21,8 +21,16 @@ import Web3 from 'web3';
 import { Bytes, ExchangeIssuanceParams, SetProtocolUtils } from 'set-protocol-utils';
 
 import { Assertions } from '../assertions';
-import { ExchangeIssuanceModuleWrapper } from '../wrappers';
-import { Address, KyberTrade, Tx, ZeroExSignedFillOrder } from '../types/common';
+import { ZERO } from '../constants';
+import { BigNumber, generateFutureTimestamp,  } from '../util';
+import {
+  ExchangeIssuanceModuleWrapper,
+  ERC20Wrapper,
+  KyberNetworkWrapper,
+  SetTokenWrapper,
+  VaultWrapper,
+ } from '../wrappers';
+import { Address, Component, KyberTrade, SetProtocolConfig, Tx, ZeroExSignedFillOrder } from '../types/common';
 import { coreAPIErrors } from '../errors';
 
 /**
@@ -37,23 +45,32 @@ export class ExchangeIssuanceAPI {
   private exchangeIssuance: ExchangeIssuanceModuleWrapper;
   private setProtocolUtils: SetProtocolUtils;
 
+  private setToken: SetTokenWrapper;
+  private erc20: ERC20Wrapper;
+  private vault: VaultWrapper;
+  private kyberNetworkWrapper: KyberNetworkWrapper;
+
   /**
    * Instantiates a new ExchangeIssuanceAPI instance that contains methods for issuing and redeeming Sets
    *
    * @param web3                         The Web3.js Provider instance you would like the SetProtocol.js library
    *                                      to use for interacting with the Ethereum network
    * @param assertions                   An instance of the Assertion library
-   * @param exchangeIssuanceAddress  The address of the ExchangeIssuanceModuleWrapper Library
+   * @param config      Configuration object conforming to SetProtocolConfig with Set Protocol's contract addresses
    */
   constructor(
     web3: Web3,
     assertions: Assertions,
-    exchangeIssuanceAddress: Address,
+    config: SetProtocolConfig,
   ) {
     this.web3 = web3;
     this.setProtocolUtils = new SetProtocolUtils(this.web3);
     this.assert = assertions;
-    this.exchangeIssuance = new ExchangeIssuanceModuleWrapper(web3, exchangeIssuanceAddress);
+    this.exchangeIssuance = new ExchangeIssuanceModuleWrapper(web3, config.exchangeIssuanceModuleAddress);
+    this.erc20 = new ERC20Wrapper(this.web3);
+    this.kyberNetworkWrapper = new KyberNetworkWrapper(this.web3, config.kyberNetworkWrapperAddress);
+    this.setToken = new SetTokenWrapper(this.web3);
+    this.vault = new VaultWrapper(this.web3, config.vaultAddress);
   }
 
   /**
@@ -78,6 +95,88 @@ export class ExchangeIssuanceAPI {
       orderData,
       txOpts,
     );
+  }
+
+  /**
+   * Generates a 256-bit salt that can be included in an order, to ensure that the order generates
+   * a unique orderHash and will not collide with other outstanding orders that are identical
+   *
+   * @return    256-bit number that can be used as a salt
+   */
+  public generateSalt(): BigNumber {
+    return SetProtocolUtils.generateSalt();
+  }
+
+  /**
+   * Generates a timestamp represented as seconds since unix epoch. The timestamp is intended to be
+   * used to generate the expiration of an issuance order
+   *
+   * @param  seconds    Seconds from the present time
+   * @return            Unix timestamp (in seconds since unix epoch)
+   */
+  public generateExpirationTimestamp(seconds: number): BigNumber {
+    return generateFutureTimestamp(seconds);
+  }
+
+  /**
+   * Calculates additional amounts of each component token in a Set needed in order to issue a specific quantity of
+   * the Set. This includes token balances a user may have in both the account wallet and the Vault contract. Can be
+   * used as `requiredComponents` and `requiredComponentAmounts` inputs for an issuance order
+   *
+   * @param  setAddress       Address of the Set token for issuance order
+   * @param  makerAddress     Address of user making the issuance order
+   * @param  quantity         Amount of the Set token to create as part of issuance order
+   * @return                  List of objects conforming to the `Component` interface with address and units of each
+   *                            component required for issuance
+   */
+  public async calculateRequiredComponentsAndUnitsAsync(
+    setAddress: Address,
+    makerAddress: Address,
+    quantity: BigNumber,
+  ): Promise<Component[]> {
+    const components = await this.setToken.getComponents(setAddress);
+    const componentUnits = await this.setToken.getUnits(setAddress);
+    const naturalUnit = await this.setToken.naturalUnit(setAddress);
+    const totalUnitsNeeded = _.map(componentUnits, componentUnit => componentUnit.mul(quantity).div(naturalUnit));
+
+    const requiredComponents: Component[] = [];
+
+    // Gather how many components are owned by the user in balance/vault
+    await Promise.all(
+      components.map(async (componentAddress, index) => {
+        const walletBalance = await this.erc20.balanceOf(componentAddress, makerAddress);
+        const vaultBalance = await this.vault.getBalanceInVault(componentAddress, makerAddress);
+        const userTokenbalance = walletBalance.add(vaultBalance);
+
+        const missingUnits = totalUnitsNeeded[index].sub(userTokenbalance);
+        if (missingUnits.gt(ZERO)) {
+          const requiredComponent: Component = {
+            address: componentAddress,
+            unit: missingUnits,
+          };
+
+          requiredComponents.push(requiredComponent);
+        }
+      }),
+    );
+
+    return requiredComponents;
+  }
+
+  /**
+   * Fetch the conversion rate for a Kyber trading pair
+   *
+   * @param  makerTokenAddress       Address of the token to trade
+   * @param  componentTokenAddress   Address of the set component to trade for
+   * @param  quantity                Quantity of maker token to trade for component token
+   * @return                         The conversion rate and slip rate for the trade
+   */
+  public async getKyberConversionRate(
+    makerTokenAddress: Address,
+    componentTokenAddress: Address,
+    quantity: BigNumber
+  ): Promise<[BigNumber, BigNumber]> {
+    return await this.kyberNetworkWrapper.conversionRate(makerTokenAddress, componentTokenAddress, quantity);
   }
