+++ src/errors/index.ts
@@ -19,7 +19,7 @@
 export { coreAPIErrors, coreAssertionErrors } from './coreErrors';
 export { erc20AssertionErrors } from './erc20Errors';
 export { exchangeIssuanceErrors } from './exchangeIssuanceErrors';
-export { orderErrors } from './orderErrors';
+export { exchangeErrors } from './exchangeErrors';
 export { rebalancingErrors } from './rebalancingErrors';
 export { schemaAssertionsError } from './schemaErrors';
 export { setTokenAssertionsErrors } from './setTokenErrors';
